Fix: 투표 - 매매 로직 연동 개선

+++ vote-service/src/main/java/com/example/vote_service/dto/payload/VoteApprovedPayloadDTO.java
@@ -11,7 +11,7 @@ public record VoteApprovedPayloadDTO(
         String historyType,       // "TRADE" | "PAY"
         String side,              // "BUY" | "SELL" | "PAY"
         String stockName,
-        Integer shares,
+        Float shares,             // 소수점 거래 지원을 위해 Float으로 변경
         Integer unitPrice,
         String currency           // "KRW" | "USD"
 ) {
